requirements.txt
python-telegram-bot==13.15
requests
urllib3==1.26.18